Change directories of policies
